# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TKOH OMS Backend Server",
        "title": "TKOH OMS",
        "contact": {
            "name": "So Nim Wai"
        },
        "license": {
            "name": "LSCM"
        },
        "version": "0.1"
    },
    "host": "20.239.95.146",
    "basePath": "/oms/",
    "paths": {
...
        "/getDutyRooms": {
            "get": {
                "description": "Get the list of location.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map Handling"
                ],
                "summary": "Get Duty Rooms.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mapHandling.GetDutyRoomsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
        "/getFloorPlan": {
            "get": {
                "description": "Get UI Floor Plan.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map Handling"
                ],
                "summary": "Get Floor Plan.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mapHandling.GetFloorPlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
...
        "/loginAdmin": {
            "post": {
                "description": "Login to OMS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login Auth"
                ],
                "summary": "Login to OMS.",
                "parameters": [
                    {
                        "description": "Login Parameters",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginAdminDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginAuth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
        "/loginStaff": {
            "post": {
                "description": "Login to OMS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login Auth"
                ],
                "summary": "Login to OMS.",
                "parameters": [
                    {
                        "description": "Login Parameters",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginStaffDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginAuth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Logout from OMS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login Auth"
                ],
                "summary": "Logout from OMS.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginAuth.LogoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
        "/renewToken": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Using Valid Token to renew token before expired",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login Auth"
                ],
                "summary": "Renew JWT Token.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginAuth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.FailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.LoginAdminDTO": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.LoginStaffDTO": {
            "type": "object",
            "properties": {
                "dutyLocationId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "loginAuth.LoginResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/loginAuth.LoginResponseBody"
                },
                "header": {
                    "$ref": "#/definitions/models.ResponseHeader"
                }
            }
        },
        "loginAuth.LoginResponseBody": {
            "type": "object",
            "properties": {
                "authToken": {
                    "type": "string"
                },
                "dutyLocationId": {
                    "type": "integer"
                },
                "dutyLocationName": {
                    "type": "string"
                },
                "loginDateTime": {
                    "type": "string"
                },
                "tokenExpiryDateTime": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "userType": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "loginAuth.LogoutResponse": {
            "type": "object",
            "properties": {
                "header": {
                    "$ref": "#/definitions/models.ResponseHeader"
                }
            }
        },
        "mapHandling.GetDutyRoomsResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/mapHandling.LocationListBody"
                },
                "header": {
                    "$ref": "#/definitions/models.ResponseHeader"
                }
            }
        },
        "mapHandling.GetFloorPlanResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/mapHandling.MapListBody"
                },
                "header": {
                    "$ref": "#/definitions/models.ResponseHeader"
                }
            }
        },
        "mapHandling.LocationListBody": {
            "type": "object",
            "properties": {
                "locationList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "locationId": {
                                "type": "integer"
                            },
                            "locationName": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "mapHandling.MapListBody": {
            "type": "object",
            "properties": {
                "mapList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "floorId": {
                                "type": "integer"
                            },
                            "floorImage": {
                                "type": "string"
                            },
                            "floorName": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "models.FailResponse": {
            "type": "object",
            "properties": {
                "header": {
                    "$ref": "#/definitions/models.FailResponseHeader"
                }
            }
        },
        "models.FailResponseHeader": {
            "type": "object",
            "properties": {
                "failedReason": {
                    "type": "string"
                },
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "models.ResponseHeader": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "integer"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}